requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
awscli==1.25.83
backcall==0.2.0
billiard==3.6.4.0
boto3==1.24.82
botocore==1.27.82
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
decorator==5.1.1
Deprecated==1.2.13
Django==4.1.1
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-redis==5.2.0
djangorestframework==3.14.0
docutils==0.16
executing==1.1.0
ipython==8.5.0
jedi==0.18.1
jmespath==1.0.1
kombu==5.2.4
matplotlib-inline==0.1.6
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
PyYAML==5.4.1
redis==4.3.4
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
traitlets==5.4.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
